backend/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dataclasses==0.8
Flask==2.0.2
fpdf==1.7.2
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pkg-resources==0.0.0
protobuf==3.19.3
python-dotenv==0.19.2
requests==2.27.0
soupsieve==2.3.1
typing-extensions==4.0.1
urllib3==1.26.7
uWSGI==2.0.20
Werkzeug==2.0.2
zipp==3.6.0
